Rollback from microsoft auth to classic app auth
