+++ backend/src/main/resources/application.properties
@@ -13,7 +13,7 @@ quarkus.hibernate-orm.database.generation=update
 %test.quarkus.http.port=9090
 
 # OIDC Configuration
-quarkus.oidc.auth-server-url=${KEYCLOAK_HOST:http://127.0.0.1:2604/auth}/realms/Betterfleet
+quarkus.oidc.auth-server-url=${KEYCLOAK_HOST:http://localhost:2604/auth}/realms/Betterfleet
 quarkus.oidc.client-id=${KEYCLOAK_CLIENT_ID:application}
 
 # Custom variables
