+++ deployment/.env
@@ -3,5 +3,5 @@ POSTGRES_PASSWORD=
 PUBLIC_QUARKUS_HOSTNAME=
 KEYCLOAK_PASSWORD=
 PUBLIC_KEYCLOAK_HOSTNAME=
-MICROSOFT_CLIENT_SECRET=
 MICROSOFT_CLIENT_ID=
+MICROSOFT_CLIENT_SECRET=
