+++ deployment/dev/.env
@@ -3,5 +3,5 @@ POSTGRES_PASSWORD=A7aXfgh4@BM#7x85
 KEYCLOAK_USER=admin
 KEYCLOAK_PASSWORD=pa$$w0rd
 PUBLIC_KEYCLOAK_HOSTNAME=http://127.0.0.1:2604
-MICROSOFT_CLIENT_SECRET=123
 MICROSOFT_CLIENT_ID=123
+MICROSOFT_CLIENT_SECRET=123
