+++ webapp/src/assets/locales/fr.json
@@ -131,9 +131,9 @@
   },
   "login": {
     "continue": "Continuer",
-    "description": "Bienvenue sur Better Fleet ! Nous sommes ravis de vous compter parmi nous. Pour que vous puissiez utiliser pleinement l'application, il est nécessaire de connecter votre compte Microsoft à celle-ci.",
+    "description": "Bienvenue sur Better Fleet ! Nous sommes ravis de vous compter parmi nous. Pour que vous puissiez utiliser pleinement l'application, il est nécessaire de vous connecter à votre compte ou bien en créer un !",
     "disconnect": "Ce n’est pas votre compte ? ",
-    "loginButton": "Se connecter à XBOX",
+    "loginButton": "Se connecter",
     "succeed": "Connexion réussie !",
     "userWelcome": "Bienvenue",
     "welcome": "Bienvenue sur"
