+++ webapp/src/objects/fleet/WebSocet.ts
@@ -13,5 +13,6 @@ export enum WebSocketMessageType {
   LEAVE_SERVER = "LEAVE_SERVER",
   OUTDATED_CLIENT = "OUTDATED_CLIENT",
   SESSION_NOT_FOUND = "SESSION_NOT_FOUND",
-  KEEP_ALIVE = "KEEP_ALIVE"
+  KEEP_ALIVE = "KEEP_ALIVE",
+  CONNECTION_REFUSED= "CONNECTION_REFUSED",
 }
