+++ webapp/src/router/index.ts
@@ -1,13 +1,14 @@
 import {createWebHistory, createRouter} from "vue-router";
-import Home from "@/components/Home.vue";
+import Home from "@/components/fleet/Home.vue";
 import fleet from "@/assets/icons/navigation.svg"
 import config from "@/assets/icons/config.svg"
 import sot from "@/assets/icons/logo.svg"
-import Fleet from "@/components/Fleet.vue";
-import Config from "@/components/Config.vue";
+import Fleet from "@/components/fleet/Fleet.vue";
+import Config from "@/components/fleet/Config.vue";
 import i18n from "@/objects/i18n";
 import {keycloakStore} from "@/objects/stores/LoginStates.ts";
 import Authentication from "@/components/Authentication.vue";
+import FleetMenuNavigator from "@/components/FleetMenuNavigator.vue";
 
 declare module 'vue-router' {
   interface RouteMeta {
@@ -24,17 +25,6 @@ const {t} = i18n.global;
 export const routes = [
   {
     path: "/",
-    name: "Home",
-    component: Home,
-    meta: {
-      icon: sot,
-      tooltip: t('tooltips.navbar.home'),
-      requiresAuth: true,
-      displayInNav: true
-    }
-  },
-  {
-    path: "/auth",
     name: "Auth",
     component: Authentication,
     meta: {
@@ -44,25 +34,47 @@ export const routes = [
   },
   {
     path: "/fleet",
-    name: "Fleet",
-    component: Fleet,
+    name: "FleetManager",
+    component: FleetMenuNavigator,
     meta: {
-      icon: fleet,
-      tooltip: t('tooltips.navbar.fleet'),
-      requiresAuth: true,
-      displayInNav: true
-    }
-  }, {
-    path: "/config",
-    name: "Config",
-    component: Config,
-    meta: {
-      icon: config,
-      tooltip: t('tooltips.navbar.config'),
-      requiresAuth: true,
-      displayInNav: true
-    }
-  },
+      displayInNav: false
+    },
+    children: [
+      {
+        path: "home",
+        name: "Home",
+        component: Home,
+        meta: {
+          icon: sot,
+          tooltip: t('tooltips.navbar.home'),
+          requiresAuth: true,
+          displayInNav: true
+        }
+      },
+      {
+        path: "session",
+        name: "Fleet",
+        component: Fleet,
+        meta: {
+          icon: fleet,
+          tooltip: t('tooltips.navbar.fleet'),
+          requiresAuth: true,
+          displayInNav: true
+        }
+      },
+      {
+        path: "config",
+        name: "Config",
+        component: Config,
+        meta: {
+          icon: config,
+          tooltip: t('tooltips.navbar.config'),
+          requiresAuth: true,
+          displayInNav: true
+        }
+      }
+    ]
+  }
 ];
 
 export const router = createRouter({
